api/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
genai-processors
dataclasses-json
pydantic
